Source "Most Popular" list from /popular
On the homepage the "most popular" list is being populated with an AJAX request
which makes scraping it difficult. On the "/popular" route, however, this is not
the case. One caveat here is that that page lists the 30 most popular stories.
Keeping with the original theme, only the first 5 are returned.